feat: refactor radar resource and remove RadarView class
