+++ src/main/java/ch/sobrado/radar/Position.java
@@ -1,0 +1,4 @@
+package ch.sobrado.radar;
+
+public enum Position {
+}
