+++ src/main/java/ch/sobrado/radar/RadarResource.java
@@ -1,25 +1,69 @@
 package ch.sobrado.radar;
 
-import jakarta.ws.rs.GET;
-import jakarta.ws.rs.Path;
-import jakarta.ws.rs.PathParam;
-import jakarta.ws.rs.Produces;
-import jakarta.ws.rs.core.MediaType;
+import java.util.Collections;
+import java.util.List;
 
-@Path("/radar")
-public class RadarResource {
+import jakarta.transaction.Transactional;
+import jakarta.ws.rs.*;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
+
+@Path("/radars")
+@Produces(MediaType.APPLICATION_JSON)
+@Consumes(MediaType.APPLICATION_JSON)
+public class RadarsResource {
 
     @GET
-    @Produces(MediaType.TEXT_PLAIN)
-    public String hello() {
-        return "Hello from RESTEasy on Sobrado Radar";
+    public List<Radar> getAll() {
+        return Radar.listAll();
     }
 
     @GET
-    @Produces(MediaType.APPLICATION_JSON)
-    @Path("/data/{year}")
-    public String data(@PathParam("year") int year) {
-        RadarView e = RadarView.find("year", year).firstResult();
-        return e.jsondata;
+    @Path("/{radarId}")
+    public Response getById(@PathParam("radarId") Long radarId) {
+        Radar radar = Radar.findById(radarId);
+        if (radar == null) {
+            return Response.status(Response.Status.NOT_FOUND).build();
+        }
+        return Response.ok(radar).build();
+    }
+
+    @POST
+    @Transactional
+    public Response create(Radar radar) {
+        radar.persist();
+        return Response.status(Response.Status.CREATED)
+                .entity(Collections.singletonMap("id", radar.id))
+                .build();
+    }
+
+    @PUT
+    @Path("/{radarId}")
+    @Transactional
+    public Response update(@PathParam("radarId") Long radarId,
+                           Radar updatedRadar) {
+
+        Radar existingRadar = Radar.findById(radarId);
+        if (existingRadar == null) {
+            return Response.status(Response.Status.NOT_FOUND).build();
+        }
+
+        existingRadar.name = updatedRadar.name;
+
+        return Response.ok(existingRadar).build();
+    }
+
+    @DELETE
+    @Path("/{radarId}")
+    @Transactional
+    public Response delete(@PathParam("radarId") Long radarId) {
+
+        Radar radar = Radar.findById(radarId);
+        if (radar == null) {
+            return Response.status(Response.Status.NOT_FOUND).build();
+        }
+
+        radar.delete();
+        return Response.noContent().build();
     }
 }
